refactor(atlas-kernels): split lib.rs tests out — the 500-LoC guard is red on main
`crates/atlas-kernels/src/lib.rs` went 496 -> 517 LoC in dd2ac46d5 (#389, the
closure-hash work — `TARGET_CLOSURES` and its doc comment). That is over the
repo's cap and NOT in `file-size-cap.yml`'s allow-list, so the guard has been
failing on main ever since, and therefore red on every PR opened since. Mine
to fix.

Carried here rather than in the kernel-compile PR so this branch — the
bootstrap that unblocks the gate — is green on everything except its own
benchmark gate. Two blockers, one merge.

The fix is the minimum that clears it and adds no indirection: the `mod
tests` block moves to a sibling `lib_tests.rs`, re-attached with `#[path]`.
That is the idiom already used by `atlas-closure` and `atlas-plugin::gate`,
and it is an exact move — all five tests still collect and run (1 passing +
4 `#[ignore]`d, unchanged).

517 -> 447 LoC. No allow-list entry added; a carve-out for a file that only
needed its tests moved would be debt for nothing.
